refactor(transfer): decompose generator/mod.rs into per-concern submodules
Split the 1027-line generator/mod.rs into focused submodules under
crates/transfer/src/generator/, leaving mod.rs as the module entry that
re-exports the public API:

- context.rs: GeneratorContext struct and inherent methods
- diagnostics.rs: process-wide INC_RECURSE counters and recorders
- io_error_flags.rs: IOERR_* bitfield constants and to_exit_code mapper
- segments.rs: PendingSegment, DirSegment, TaggedIndex, SegmentScheduler,
  IncrementalState, MIN_FILECNT_LOOKAHEAD
- stats.rs: GeneratorStats, TransferLoopResult, is_early_close_error
- timing.rs: TransferTiming

Public re-exports (GeneratorContext, GeneratorStats,
generate_delta_from_signature, io_error_flags, ItemFlags,
calculate_duration_ms, read_signature_blocks, flush_rate_totals,
ndx_convert_totals, prepare_acl_totals, segment_dispatch_totals) are
preserved byte-for-byte through mod.rs.
